requirements.txt
audioread==3.0.1
av==12.0.0
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
decorator==4.4.2
distlib==0.3.8
filelock==3.13.4
identify==2.5.35
idna==3.7
imageio==2.34.0
imageio-ffmpeg==0.4.9
joblib==1.4.0
lazy-loader==0.4
librosa==0.10.1
llvmlite==0.42.0
moviepy==1.0.3
msgpack==1.0.8
nodeenv==1.8.0
numba==0.59.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.0
pooch==1.8.1
pre-commit==3.7.0
proglog==0.1.10
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2024.1
pyyaml==6.0.1
requests==2.31.0
ruff==0.3.7
scikit-learn==1.4.2
scipy==1.13.0
setuptools==69.2.0
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
threadpoolctl==3.4.0
tqdm==4.66.2
typing-extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.1
cfgv==3.4.0
distlib==0.3.8
filelock==3.13.4
identify==2.5.35
nodeenv==1.8.0
platformdirs==4.2.0
pre-commit==3.7.0
pyyaml==6.0.1
ruff==0.3.7
setuptools==69.2.0
virtualenv==20.25.1
psycopg2-binary==2.9.9
numpy==1.26.4
pgvector==0.2.5
pandas==2.2.2
python-vlc==3.0.20123
PyWavelets==1.6.0
Flask==3.0.3